Modularized code to improve readability
